Fix 128-byte limit on data received on serial port
The solution is to wait for all data we want to receive as part of a
command.

Also removed the delay which was added before the call to RS_read() in
obcsim.ino, the one which was thought to fix the problem.

The problem actually seems to arise because of the Arduino's limited
buffer size. The definition for the buffer size is in Arduino's
HardwareSerial.h file, and it is:

(line 53) #define SERIAL_RX_BUFFER_SIZE 64

Therefore, 64 bytes wide. This setting seems to however be changed
somewhere (I couldn't figure out where) in the Arduino IDE, for the
Arduino Due we use.