Re-send duplicate data frame
Because of the way MSP handles T_ACKs, the possibility existed for
configuration data not being sent correctly to the experiment.
Essentially, if the sequence flags got de-synced because of an Arduino
reboot, configuration data sent with a de-synced sequence flag was
replied to by the experiment with a T_ACK to signal a duplicate frame.
This meant that the configuration data was not applied, but the OBC was
thinking the data was sent correctly, when it actually wasn't.

This was fixed by:

1. Moving the incrementation of the MSP link's processed_length member
to the msp_send_data_frame() function;

2. Adding a new type of error to msp_obc_error.h, to signal that not all
data was sent (MSP_OBC_ERR_DATA_NOT_SENT;

3. Checking in msp_recv_header_frame(), when a T_ACK is received from
the experiment, that all data bytes were actually sent. This is done by
checking that processed_length == total_length;

4. Returning from msp_recv_header_frame() with the new type of error;

5. Re-issuing the transaction if this error is encountered.

Note that with the current way of re-issuing the transaction, the
possibbility exists for a race condition, whereby invoke_send() is
always called. The current implementation covers de-synced TIDs, but
there may be cases it doesn't cover.

The TODO on top of te recursive call to invoke_send() suggests a way of
fixing this.