Move single request buffer allocation to RS232.cpp
This makes the request buffer actually be usable outside of
obcsim_transaction.cpp. With the way it was previously implemented, the
receive buffer pointer given as an argument to invoke_request() was
being assigned to the request buffer inside obcsim_transactions.cpp, in
the hope that this assignment would set the external pointer to the
request buffer inside obcsim_transactions.cpp.

The fact of the matter is that C interprets the pointer offered as an
argument as an argument passed by value, so whatever assignment is made
to the pointer inside the invoke_request() function is lost when
invoke_request() exits.

The fix to this is to move the buffer definition in the file which calls
invoke_request(), make the request buffer a simple pointer inside
obcsim_transactions.cpp, and assign the internal request buffer to the
value of the pointer given as an argument to invoke_request().

This commit also creates a new buffer for repeated send requests.