Attempt to compress timed DAQ logic (WIP)
NB: WIP, as the code is untested - only compiles correctly.

This commit makes the following changes:

1. Move the timed DAQ logic from RS232 to RTC. This seems to be the
most logical place for it for now, as the RTC file is where the DAQ_DUR
value is kept.

2. Apply these changes in RS232.cpp.

3. Remove the "still_running" part in obcsim.ino. This should not be
needed, as the code stalls until the next DAQ_DUR when a DAQ_STOP
command is received inside RS232.cpp.