Read last DAQ file name from dedicated file on SD card
The name of the last DAQ file to be written is stored on the SD card in
a dedicated file called "lastfile.txt".

This commit implements the reading of this file in the daq_init()
function, which replaces the previous SD_init(). Should lastfile.txt not
exist, it is created and populated with the default DAQ file name,
"daq00000.dat".