Refactor daq_write_new_file
The code does not work yet as intended, due to the fact that all the SD
card functions write from the end of the file. The intent with
"lastfile.txt" is to always write the last file at the first line, then
read it and overwrite it. Since the SD card functions all write to the
last position in the file, this does not work.