Re-refactor daq_init and daq_write_new_file
This time, a static variable is used in the DAQ file (currently
SDcard.cpp) to keep the file name. The currently available files on the
SD card are found on daq_init() by looking for the pattern "daqNNNN.dat"
and incrementing the NNNN number field from 0000 until a file is not
found. When it is not found, we are ready to start operation.

NB: obcsim.ino only checks the code, it does not allow for any commands
from the CitirocUI in this commit!