Refactor daq_read_last_file
The function now writes the data from the last file into a buffer passed
as a parameter, which is then to be used externally as the user desires.
In this case, the user desires to write it over RS-232.

The choice to do it like this is because we want to make the code more
readable, and not have a function do too many things in the background.

NB: Code untested.