Remove redundant and buggy clearing of recv_buf
It was redundant because the recv_buf is always (at the moment, at
least) always read immediately after it being written to during an MSP
transaction.

It was buggy because it was only writing four bytes, due to the memset's
fourth argument being sizeof(recv_len), with recv_len a four-byte
variable. Hence, not all of the recv_buf was being cleared.