Send DAQ time to CUBES in case of reset
The DAQ time also needs to be sent, as a reset will have left CUBES with
no DAQ time set. A DAQ_START command sent to CUBES will probably result
in a 255-s DAQ run on the CUBES side, due to the gateware
implementation. The 255-s DAQ time on the gateware side will definitely
result in undesired behaviour.

To be able to set the DAQ time, a new function rtc_get_daq_time() needed
to be implemented in RTC.cpp. Just so we are consistent, the
time_req_payload variable in RTC.cpp was renamed to daq_time, and the
function used to set the Arduino daq_time was renamed to
rtc_set_daq_time().