Add CUBES time sync via MSP_SEND_TIME
It turns out the on-chip clock of CUBES is quite crappy, potentially
losing one second every 100s (it has a 1% accuracy). In order to avoid
desynchronizations in time between the Arduino and CUBES and
potentially lose data, the solution adopted is (currently) to
_synchronize the time between the Arduino and CUBES once every 30 s_.

This is done using the `arduino-timer` library, which assigns a "task" to
execute every 30 s; the "task" sends the MSP_SEND_TIME command to CUBES
to set the new time.