Change DAQ configuration and some other MSP cmds
Two bytes are now sent as part of the DAQ configuration, the DAQ
duration and the bin configuration. Changes in both the UART and MSP
commands have been made to reflect this.

In addition, the unused HVPS_ON/OFF MSP commands have been removed, and
the op-codes of the DAQ_START and _STOP commands have been "promoted".