Migrate DAQ implementation and clean up RTC files
The RTC files were previously used as a mix between controlling the RTC
and implementing the DAQ timer. It for long made more sense to me to
have the DAQ timer stuff in the DAQ files (`daq.cpp`, `daq.hpp`).

This commit implements the migration of DAQ timer to the DAQ files.
Functions that used to be in the RTC files were renamed accordingly.

I've also taken the opportunity to clean up the RTC functions, unused
functions were removed and the remaining ones were renamed to be as
simple as possible.