Ensure debug print code works
The sizes of some temporary strings to be used with `sprintf()` have
been increased, to ensure no "segfaults" occur on the Arduino (one has
been noticed!). In addition, the extra line in between "response:
BLABLA" and the "Sent/Received bytes" has been removed on MSP commands.
This makes the printout more compact - and consistent.

"Ensure" in the commit title is to some extent, of course. There may
be untested cases which do not work properly.