Update sync logic
The Arduino now only syncs its time with CUBES once every 30 seconds
when DAQ is on-going. Instead of using an Arduino timer "object" from
the library, the local RTC timer is used as for the DAQ duration
"timer". The value of the RTC timer is compared in another function
inside the main Arduino `loop()` to trigger the sync every 30 seconds.